Fixed casting warnings when comparing MAX_DIST.
  deflate_medium.c(127,76): warning C4244: '=': conversion from 'unsigned int' to 'Pos', possible loss of data

+++ deflate_p.h
@@ -45,8 +45,8 @@ static inline int zng_tr_tally_dist(deflate_state *s, uint32_t dist, uint32_t le
     s->sym_buf[s->sym_next++] = (uint8_t)len;
     s->matches++;
     dist--;
-    Assert((uint16_t)dist < (uint16_t)MAX_DIST(s) &&
-           (uint16_t)d_code(dist) < (uint16_t)D_CODES,  "zng_tr_tally: bad match");
+    Assert(dist < MAX_DIST(s) && (uint16_t)d_code(dist) < (uint16_t)D_CODES, 
+        "zng_tr_tally: bad match");
 
     s->dyn_ltree[zng_length_code[len]+LITERALS+1].Freq++;
     s->dyn_dtree[d_code(dist)].Freq++;
